+++ tools/data-handler/test/mutations/resource-mutations.test.ts
@@ -3,15 +3,15 @@ import { mkdir } from 'node:fs/promises';
 import { join } from 'node:path';
 
 import { Project } from '../../src/containers/project.js';
-import { ResourceMutations } from '../../src/mutations/plan.js';
+import { ResourceMutations } from '../../src/mutations/resource-mutations.js';
 import { resourceName } from '../../src/utils/resource-utils.js';
 import { copyDir, deleteDir } from '../../src/utils/file-utils.js';
 import { ConfigurationLogger } from '../../src/utils/configuration-logger.js';
 
-const testDir = join(import.meta.dirname, 'tmp-plan');
+const testDir = join(import.meta.dirname, 'tmp-resource-mutations');
 const fixturePath = join(testDir, 'valid', 'decision-records');
 
-describe('ResourceMutations.plan + apply', () => {
+describe('ResourceMutations.apply', () => {
   let project: Project;
 
   beforeAll(async () => {
@@ -24,7 +24,7 @@ describe('ResourceMutations.plan + apply', () => {
     await deleteDir(testDir);
   });
 
-  it('plan() returns a PreviewResult for a display-only edit', async () => {
+  it('apply() succeeds for a non-cascading edit', async () => {
     const mutations = new ResourceMutations(project);
     const input = {
       kind: 'edit' as const,
@@ -36,25 +36,7 @@ describe('ResourceMutations.plan + apply', () => {
         to: 'New',
       },
     };
-    const result = await mutations.plan(input);
-    expect(result.isBreaking).toBe(false);
-    expect(result.preview.affectedCardCount).toBe(0);
-    expect(result.fingerprint.digest).toMatch(/^[0-9a-f]{64}$/);
-  });
-
-  it('apply() succeeds for a non-cascading edit without fingerprint', async () => {
-    const mutations = new ResourceMutations(project);
-    const input = {
-      kind: 'edit' as const,
-      target: resourceName(`${project.projectPrefix}/cardTypes/decision`),
-      updateKey: { key: 'displayName' },
-      operation: {
-        name: 'change' as const,
-        target: 'Decision card type',
-        to: 'New',
-      },
-    };
-    await expect(mutations.apply(input)).resolves.toEqual({ success: true });
+    await expect(mutations.apply(input)).resolves.toBeUndefined();
   });
 
   it('apply() with project_rename input writes a project_rename log entry', async () => {
